// src/hey-api-generated-ksef-client/types.gen.ts
export type AmountType = 'Brutto' | 'Netto' | 'Vat';
// ...
export type BuyerIdentifierType = 'Nip' | 'VatUe' | 'Other' | 'None';
// ...
export type CurrencyCode =
  | 'AED'
  | 'AFN'
  | 'ALL'
  | 'AMD'
  | 'ANG'
  | 'AOA'
  | 'ARS'
  | 'AUD'
  | 'AWG'
  | 'AZN'
  | 'BAM'
  | 'BBD'
  | 'BDT'
  | 'BGN'
  | 'BHD'
  | 'BIF'
  | 'BMD'
  | 'BND'
  | 'BOB'
  | 'BOV'
  | 'BRL'
  | 'BSD'
  | 'BTN'
  | 'BWP'
  | 'BYN'
  | 'BZD'
  | 'CAD'
  | 'CDF'
  | 'CHE'
  | 'CHF'
  | 'CHW'
  | 'CLF'
  | 'CLP'
  | 'CNY'
  | 'COP'
  | 'COU'
  | 'CRC'
  | 'CUC'
  | 'CUP'
  | 'CVE'
  | 'CZK'
  | 'DJF'
  | 'DKK'
  | 'DOP'
  | 'DZD'
  | 'EGP'
  | 'ERN'
  | 'ETB'
  | 'EUR'
  | 'FJD'
  | 'FKP'
  | 'GBP'
  | 'GEL'
  | 'GGP'
  | 'GHS'
  | 'GIP'
  | 'GMD'
  | 'GNF'
  | 'GTQ'
  | 'GYD'
  | 'HKD'
  | 'HNL'
  | 'HRK'
  | 'HTG'
  | 'HUF'
  | 'IDR'
  | 'ILS'
  | 'IMP'
  | 'INR'
  | 'IQD'
  | 'IRR'
  | 'ISK'
  | 'JEP'
  | 'JMD'
  | 'JOD'
  | 'JPY'
  | 'KES'
  | 'KGS'
  | 'KHR'
  | 'KMF'
  | 'KPW'
  | 'KRW'
  | 'KWD'
  | 'KYD'
  | 'KZT'
  | 'LAK'
  | 'LBP'
  | 'LKR'
  | 'LRD'
  | 'LSL'
  | 'LYD'
  | 'MAD'
  | 'MDL'
  | 'MGA'
  | 'MKD'
  | 'MMK'
  | 'MNT'
  | 'MOP'
  | 'MRU'
  | 'MUR'
  | 'MVR'
  | 'MWK'
  | 'MXN'
  | 'MXV'
  | 'MYR'
  | 'MZN'
  | 'NAD'
  | 'NGN'
  | 'NIO'
  | 'NOK'
  | 'NPR'
  | 'NZD'
  | 'OMR'
  | 'PAB'
  | 'PEN'
  | 'PGK'
  | 'PHP'
  | 'PKR'
  | 'PLN'
  | 'PYG'
  | 'QAR'
  | 'RON'
  | 'RSD'
  | 'RUB'
  | 'RWF'
  | 'SAR'
  | 'SBD'
  | 'SCR'
  | 'SDG'
  | 'SEK'
  | 'SGD'
  | 'SHP'
  | 'SLL'
  | 'SOS'
  | 'SRD'
  | 'SSP'
  | 'STN'
  | 'SVC'
  | 'SYP'
  | 'SZL'
  | 'THB'
  | 'TJS'
  | 'TMT'
  | 'TND'
  | 'TOP'
  | 'TRY'
  | 'TTD'
  | 'TWD'
  | 'TZS'
  | 'UAH'
  | 'UGX'
  | 'USD'
  | 'USN'
  | 'UYI'
  | 'UYU'
  | 'UYW'
  | 'UZS'
  | 'VES'
  | 'VND'
  | 'VUV'
  | 'WST'
  | 'XAF'
  | 'XAG'
  | 'XAU'
  | 'XBA'
  | 'XBB'
  | 'XBC'
  | 'XBD'
  | 'XCD'
  | 'XCG'
  | 'XDR'
  | 'XOF'
  | 'XPD'
  | 'XPF'
  | 'XPT'
  | 'XSU'
  | 'XUA'
  | 'XXX'
  | 'YER'
  | 'ZAR'
  | 'ZMW'
  | 'ZWL';
// ...
export type FormCode = {
  /**
   * Kod systemowy
   */
  systemCode: string;
  /**
   * Wersja schematu
   */
  schemaVersion: string;
  /**
   * Wartość
   */
  value: string;
};
// ...
export type InvoiceMetadata = {
  /**
   * Numer KSeF faktury.
   */
  ksefNumber: KsefNumber;
  /**
   * Numer faktury nadany przez wystawcę.
   */
  invoiceNumber: string;
  /**
   * Data wystawienia faktury.
   */
  issueDate: string;
  /**
   * Data przyjęcia faktury w systemie KSeF (do dalszego przetwarzania).
   */
  invoicingDate: string;
  /**
   * Data nadania numeru KSeF.
   */
  acquisitionDate: string;
  /**
   * Data trwałego zapisu faktury w repozytorium systemu KSeF.
   */
  permanentStorageDate: string;
  /**
   * Dane identyfikujące sprzedawcę.
   */
  seller: InvoiceMetadataSeller;
  /**
   * Dane identyfikujące nabywcę.
   */
  buyer: InvoiceMetadataBuyer;
  /**
   * Łączna kwota netto.
   */
  netAmount: number;
  /**
   * Łączna kwota brutto.
   */
  grossAmount: number;
  /**
   * Łączna kwota VAT.
   */
  vatAmount: number;
  /**
   * Kod waluty.
   */
  currency: string;
  /**
   * Tryb fakturowania (online/offline).
   */
  invoicingMode: InvoicingMode;
  /**
   * Rodzaj faktury.
   * | Wartość | Opis |
   * | --- | --- |
   * | Vat | (FA) Podstawowa |
   * | Zal | (FA) Zaliczkowa |
   * | Kor | (FA) Korygująca |
   * | Roz | (FA) Rozliczeniowa |
   * | Upr | (FA) Uproszczona |
   * | KorZal | (FA) Korygująca fakturę zaliczkową |
   * | KorRoz | (FA) Korygująca fakturę rozliczeniową |
   * | VatPef | (PEF) Podstawowowa |
   * | VatPefSp | (PEF) Specjalizowana |
   * | KorPef | (PEF) Korygująca |
   * | VatRr | (RR) Podstawowa |
   * | KorVatRr | (RR) Korygująca |
   *
   */
  invoiceType: InvoiceType;
  /**
   * Struktura dokumentu faktury.
   *
   * Obsługiwane schematy:
   * | SystemCode | SchemaVersion | Value |
   * | --- | --- | --- |
   * | [FA (2)](https://github.com/CIRFMF/ksef-docs/blob/main/faktury/schemy/FA/schemat_FA(2)_v1-0E.xsd) | 1-0E | FA |
   * | [FA (3)](https://github.com/CIRFMF/ksef-docs/blob/main/faktury/schemy/FA/schemat_FA(3)_v1-0E.xsd) | 1-0E | FA |
   * | [PEF (3)](https://github.com/CIRFMF/ksef-docs/blob/main/faktury/schemy/PEF/Schemat_PEF(3)_v2-1.xsd) | 2-1 | PEF |
   * | [PEF_KOR (3)](https://github.com/CIRFMF/ksef-docs/blob/main/faktury/schemy/PEF/Schemat_PEF_KOR(3)_v2-1.xsd) | 2-1 | PEF |
   * | FA_RR (1) | 1-0E | RR |
   *
   */
  formCode: FormCode;
  /**
   * Czy faktura została wystawiona w trybie samofakturowania.
   */
  isSelfInvoicing: boolean;
  /**
   * Określa, czy faktura posiada załącznik.
   */
  hasAttachment: boolean;
  /**
   * Skrót SHA256 faktury.
   */
  invoiceHash: Sha256HashBase64;
  /**
   * Skrót SHA256 korygowanej faktury.
   */
  hashOfCorrectedInvoice?: Sha256HashBase64 | null;
  /**
   * Lista podmiotów trzecich.
   */
  thirdSubjects?: Array<InvoiceMetadataThirdSubject> | null;
  /**
   * Podmiot upoważniony.
   */
  authorizedSubject?: InvoiceMetadataAuthorizedSubject | null;
};
// ...
export type InvoiceMetadataBuyer = {
  /**
   * Identyfikator nabywcy.
   * | Type | Value |
   * | --- | --- |
   * | Nip | 10 cyfrowy numer NIP |
   * | VatUe | Identyfikator VAT UE podmiotu unijnego |
   * | Other | Inny identyfikator|
   * | None  | Brak identyfikatora nabywcy |
   */
  identifier: InvoiceMetadataBuyerIdentifier;
  /**
   * Nazwa nabywcy.
   */
  name?: string | null;
};
// ...
export type InvoiceMetadataSeller = {
  /**
   * Nip sprzedawcy.
   */
  nip: Nip;
  /**
   * Nazwa sprzedawcy.
   */
  name?: string | null;
};

export type InvoiceMetadataThirdSubject = {
  identifier: InvoiceMetadataThirdSubjectIdentifier;
  /**
   * Nazwa podmiotu trzeciego.
   */
  name?: string | null;
  /**
   * Rola podmiotu trzeciego.
   * | Wartość | Opis |
   * | ---- | --- |
   * | 0 | Inna rola |
   * | 1 | Faktor - w przypadku gdy na fakturze występują dane faktora |
   * | 2 | Odbiorca - w przypadku gdy na fakturze występują dane jednostek wewnętrznych, oddziałów, wyodrębnionych w ramach nabywcy, które same nie stanowią nabywcy w rozumieniu ustawy |
   * | 3 | Podmiot pierwotny - w przypadku gdy na fakturze występują dane podmiotu będącego w stosunku do podatnika podmiotem przejętym lub przekształconym, który dokonywał dostawy lub świadczył usługę. Z wyłączeniem przypadków, o których mowa w art. 106j ust.2 pkt 3 ustawy, gdy dane te wykazywane są w części Podmiot1K |
   * | 4 | Dodatkowy nabywca - w przypadku gdy na fakturze występują dane kolejnych (innych niż wymieniony w części Podmiot2) nabywców |
   * | 5 | Wystawca faktury - w przypadku gdy na fakturze występują dane podmiotu wystawiającego fakturę w imieniu podatnika. Nie dotyczy przypadku, gdy wystawcą faktury jest nabywca |
   * | 6 | Dokonujący płatności - w przypadku gdy na fakturze występują dane podmiotu regulującego zobowiązanie w miejsce nabywcy |
   * | 7 | Jednostka samorządu terytorialnego - wystawca |
   * | 8 | Jednostka samorządu terytorialnego - odbiorca |
   * | 9 | Członek grupy VAT - wystawca |
   * | 10 | Członek grupy VAT - odbiorca |
   * | 11 | Pracownik |
   */
  role: number;
};
// ...
export type InvoiceQueryAmount = {
  type: AmountType;
  from?: number | null;
  to?: number | null;
};

export type InvoiceQueryBuyerIdentifier = {
  /**
   * Typ identyfikatora nabywcy.
   * | Wartość | Opis |
   * | --- | --- |
   * | Nip | 10 cyfrowy numer NIP |
   * | VatUe | Identyfikator VAT UE podmiotu unijnego |
   * | Other | Inny identyfikator |
   * | None | Brak identyfikatora nabywcy |
   *
   */
  type: BuyerIdentifierType;
  /**
   * Wartość identyfikatora nabywcy (exact match).
   */
  value?: string | null;
};

export type InvoiceQueryDateRange = {
  /**
   * Typ daty, według której ma być zastosowany zakres.
   * | Wartość | Opis |
   * | --- | --- |
   * | Issue | Data wystawienia faktury. |
   * | Invoicing | Data przyjęcia faktury w systemie KSeF (do dalszego przetwarzania). |
   * | PermanentStorage | Data trwałego zapisu faktury w repozytorium systemu KSeF. |
   *
   */
  dateType: InvoiceQueryDateType;
  /**
   * Data początkowa zakresu(UTC).
   */
  from: string;
  /**
   * Data końcowa zakresu(UTC). Jeśli nie zostanie podana, przyjmowana jest bieżąca data i czas w UTC.
   */
  to?: string | null;
  /**
   * Określa, czy system ma ograniczyć filtrowanie (zakres dateRange.to) do wartości `PermanentStorageHwmDate`.
   *
   * * Dotyczy wyłącznie zapytań z `dateType = PermanentStorage`,
   * * Gdy `true`, system ogranicza filtrowanie tak, aby wartość `dateRange.to` nie przekraczała wartości `PermanentStorageHwmDate`,
   * * Gdy `null` lub `false`, filtrowanie może wykraczać poza `PermanentStorageHwmDate`.
   */
  restrictToPermanentStorageHwmDate?: boolean | null;
};

/**
 * | Wartość | Opis |
 * | --- | --- |
 * | Issue | Data wystawienia faktury. |
 * | Invoicing | Data przyjęcia faktury w systemie KSeF (do dalszego przetwarzania). |
 * | PermanentStorage | Data trwałego zapisu faktury w repozytorium systemu KSeF. |
 *
 */
export type InvoiceQueryDateType = 'Issue' | 'Invoicing' | 'PermanentStorage';

export type InvoiceQueryFilters = {
  /**
   * Typ podmiotu, którego dotyczą kryteria filtrowania metadanych faktur.
   * Określa kontekst, w jakim przeszukiwane są dane.
   * | Wartość | Opis |
   * | --- | --- |
   * | Subject1 | Podmiot 1 - sprzedawca |
   * | Subject2 | Podmiot 2 - nabywca |
   * | Subject3 | Podmiot 3 |
   * | SubjectAuthorized | Podmiot upoważniony |
   *
   */
  subjectType: InvoiceQuerySubjectType;
  /**
   * Typ i zakres dat, według którego mają być filtrowane faktury. Dozwolony maksymalny okres wynosi 2 lata.
   */
  dateRange: InvoiceQueryDateRange;
  /**
   * Numer KSeF faktury (exact match).
   */
  ksefNumber?: KsefNumber | null;
  /**
   * Numer faktury nadany przez wystawcę (exact match).
   */
  invoiceNumber?: string | null;
  /**
   * Filtr kwotowy – brutto, netto lub VAT (z wartością).
   */
  amount?: InvoiceQueryAmount | null;
  /**
   * Nip sprzedawcy (exact match).
   */
  sellerNip?: Nip | null;
  /**
   * Identyfikator nabywcy.
   * | Type | Value |
   * | --- | --- |
   * | Nip | 10 cyfrowy numer NIP |
   * | VatUe | Identyfikator VAT UE podmiotu unijnego. |
   * | Other | Inny identyfikator|
   * | None  | Brak identyfikatora nabywcy |
   */
  buyerIdentifier?: InvoiceQueryBuyerIdentifier | null;
  /**
   * Kody walut.
   */
  currencyCodes?: Array<CurrencyCode> | null;
  /**
   * Tryb wystawienia faktury: online lub offline.
   */
  invoicingMode?: InvoicingMode | null;
  /**
   * Czy faktura została wystawiona w trybie samofakturowania.
   */
  isSelfInvoicing?: boolean | null;
  /**
   * Typ dokumentu.
   * | Wartość | Opis |
   * | --- | --- |
   * | FA | Faktura VAT |
   * | PEF | Faktura PEF |
   * | RR | Faktura RR |
   *
   */
  formType?: InvoiceQueryFormType | null;
  /**
   * Rodzaje faktur.
   * | Wartość | Opis |
   * | --- | --- |
   * | Vat | (FA) Podstawowa |
   * | Zal | (FA) Zaliczkowa |
   * | Kor | (FA) Korygująca |
   * | Roz | (FA) Rozliczeniowa |
   * | Upr | (FA) Uproszczona |
   * | KorZal | (FA) Korygująca fakturę zaliczkową |
   * | KorRoz | (FA) Korygująca fakturę rozliczeniową |
   * | VatPef | (PEF) Podstawowowa |
   * | VatPefSp | (PEF) Specjalizowana |
   * | KorPef | (PEF) Korygująca |
   * | VatRr | (RR) Podstawowa |
   * | KorVatRr | (RR) Korygująca |
   *
   */
  invoiceTypes?: Array<InvoiceType> | null;
  /**
   * Czy faktura ma załącznik.
   */
  hasAttachment?: boolean | null;
};

/**
 * | Wartość | Opis |
 * | --- | --- |
 * | FA | Faktura VAT |
 * | PEF | Faktura PEF |
 * | RR | Faktura RR |
 *
 */
export type InvoiceQueryFormType = 'FA' | 'PEF' | 'RR';

/**
 * | Wartość | Opis |
 * | --- | --- |
 * | Subject1 | Podmiot 1 - sprzedawca |
 * | Subject2 | Podmiot 2 - nabywca |
 * | Subject3 | Podmiot 3 |
 * | SubjectAuthorized | Podmiot upoważniony |
 *
 */
export type InvoiceQuerySubjectType = 'Subject1' | 'Subject2' | 'Subject3' | 'SubjectAuthorized';
// ...
export type InvoiceType =
  | 'Vat'
  | 'Zal'
  | 'Kor'
  | 'Roz'
  | 'Upr'
  | 'KorZal'
  | 'KorRoz'
  | 'VatPef'
  | 'VatPefSp'
  | 'KorPef'
  | 'VatRr'
  | 'KorVatRr';

export type InvoicingMode = 'Online' | 'Offline';
// ...
export type KsefNumber = string;

/**
 * 10 cyfrowy numer NIP.
 */
export type Nip = string;
// ...
export type Sha256HashBase64 = string;